fix: use jest to do version test
