+++ package.json
@@ -26,7 +26,7 @@
     "test:server-protocols": "yarn build:server-protocols && lerna run test --scope '@aws-sdk/*-server'",
     "pretest:e2e": "yarn build:crypto-dependencies && lerna run --scope '@aws-sdk/{client-cloudformation,karma-credential-loader}' --include-dependencies build",
     "test:e2e": "node ./tests/e2e/index.js",
-    "test:versions": "node tests/versions/tslib.js",
+    "test:versions": "jest --config tests/versions/jest.config.js tests/versions/index.spec.ts",
     "local-publish": "node ./scripts/verdaccio-publish/index.js"
   },
   "repository": {
